Predictive discriminant analyses based on balanced samples.

More about it

pldam(mat = mat, group = group, mattoid = mattoid, codetoid = codetoid,
  ncpts = 20, limproba = 0, limCVP = FALSE, nrep = 100,
  return.matrix = TRUE, minInd = NULL)

Arguments

mat
matrix with the 'known' data (usually PC scores, but works also with univariate data)

group
grouping factor of the 'known' data

mattoid
matrix with the data to be predicted

codetoid
label of the specimens to predict

ncpts
number of PCs to retain

limproba
a limit for the posterior probability to be concidered

limCVP
a limit for the cross validation percent of the discriminate analysis to be concidered

nrep
number of replicates

return.matrix
logical whether to return full results

minInd
the number of specimens to include for the balanced design if smaller than the smallest of the "known" groups

Value

a list

Examples

# Some examples
# Grab some data from pig
mat <- pig$mat
gp <- pig$gp
# For the sake of reproducibility
set.seed(123)
train_ids <- sample(nrow(mat), 150, replace=FALSE)

pldam(# train data
      mat=mat[train_ids, ], group=gp[train_ids],
      # test data
      mattoid=mat[-train_ids, ], codetoid=paste0("id_", 1:nrow(mat[-train_ids, ])),
      nrep=2)

#> [1] "The analyses are done with 2 groups"
#> group
#>  DP  WB 
#>  36 114 
#> [1] "2  ressampling"
#> [1] "The limit % to concider an identification is  0"
#> [1] "Results are provided for the 1 LDAs that have a CVP above  0.854166666666667"

#> $CVbalanced
#> [1] 0.8333333 0.8611111
#> 
#> $resultplda
#>       rep1 rep2
#> id_1  "WB" "WB"
#> id_2  "WB" "WB"
#> id_3  "WB" "DP"
#> id_4  "WB" "WB"
#> id_5  "WB" "WB"
#> id_6  "WB" "WB"
#> id_7  "WB" "WB"
#> id_8  "WB" "WB"
#> id_9  "WB" "WB"
#> id_10 "DP" "DP"
#> id_11 "WB" "WB"
#> id_12 "WB" "WB"
#> id_13 "WB" "WB"
#> id_14 "WB" "WB"
#> id_15 "WB" "WB"
#> id_16 "DP" "DP"
#> id_17 "DP" "DP"
#> id_18 "DP" "DP"
#> id_19 "DP" "DP"
#> id_20 "DP" "DP"
#> id_21 "DP" "DP"
#> 
#> $resultpldap
#>       rep1   rep2  
#> id_1  "0.99" "1"   
#> id_2  "0.92" "1"   
#> id_3  "0.63" "0.73"
#> id_4  "1"    "1"   
#> id_5  "0.95" "0.69"
#> id_6  "0.9"  "0.99"
#> id_7  "1"    "1"   
#> id_8  "1"    "1"   
#> id_9  "0.99" "1"   
#> id_10 "0.77" "0.87"
#> id_11 "1"    "0.99"
#> id_12 "0.9"  "0.62"
#> id_13 "1"    "1"   
#> id_14 "1"    "1"   
#> id_15 "0.96" "0.99"
#> id_16 "0.99" "1"   
#> id_17 "1"    "1"   
#> id_18 "0.91" "1"   
#> id_19 "0.98" "1"   
#> id_20 "0.92" "0.86"
#> id_21 "0.88" "1"   
#> 
#> $matplda
#>       codetoid NumberOfLda ID-1 Proba-1 ID-2 Proba-2
#>  [1,] "id_1"   "1"         "WB" "100"   NA   NA     
#>  [2,] "id_2"   "1"         "WB" "100"   NA   NA     
#>  [3,] "id_3"   "1"         "DP" "100"   NA   NA     
#>  [4,] "id_4"   "1"         "WB" "100"   NA   NA     
#>  [5,] "id_5"   "1"         "WB" "100"   NA   NA     
#>  [6,] "id_6"   "1"         "WB" "100"   NA   NA     
#>  [7,] "id_7"   "1"         "WB" "100"   NA   NA     
#>  [8,] "id_8"   "1"         "WB" "100"   NA   NA     
#>  [9,] "id_9"   "1"         "WB" "100"   NA   NA     
#> [10,] "id_10"  "1"         "DP" "100"   NA   NA     
#> [11,] "id_11"  "1"         "WB" "100"   NA   NA     
#> [12,] "id_12"  "1"         "WB" "100"   NA   NA     
#> [13,] "id_13"  "1"         "WB" "100"   NA   NA     
#> [14,] "id_14"  "1"         "WB" "100"   NA   NA     
#> [15,] "id_15"  "1"         "WB" "100"   NA   NA     
#> [16,] "id_16"  "1"         "DP" "100"   NA   NA     
#> [17,] "id_17"  "1"         "DP" "100"   NA   NA     
#> [18,] "id_18"  "1"         "DP" "100"   NA   NA     
#> [19,] "id_19"  "1"         "DP" "100"   NA   NA     
#> [20,] "id_20"  "1"         "DP" "100"   NA   NA     
#> [21,] "id_21"  "1"         "DP" "100"   NA   NA     
#>